Use importer for category list handler
